+++ test/unit/mongo_rep_admin/unit_test_run.sh
@@ -19,7 +19,6 @@ echo "Unit test:"
 /usr/bin/python test/unit/mongo_rep_admin/help_message.py
 /usr/bin/python test/unit/mongo_rep_admin/main.py
 /usr/bin/python test/unit/mongo_rep_admin/node_chk.py
-/usr/bin/python test/unit/mongo_rep_admin/prt_rep_stat.py
 /usr/bin/python test/unit/mongo_rep_admin/rep_health_chk.py
 /usr/bin/python test/unit/mongo_rep_admin/rep_msg_chk.py
 /usr/bin/python test/unit/mongo_rep_admin/rep_state_chk.py