+++ test/unit/sonarqube_code_coverage.sh
@@ -21,7 +21,6 @@ coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/get_optimedat
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/help_message.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/main.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/node_chk.py
-coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/prt_rep_stat.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/rep_health_chk.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/rep_msg_chk.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/rep_state_chk.py